target_grid
Added the target_grid, and made unit spawning and checking much sleeker
using this. People can't move over people or mountains or oceans now.
Damage and durability levels tweaked for testing.

Problem that arose at time of commit: the moveCheck for moving has
problems when moving to the far end of the map, probably because things
index differently, make sure everything is all starting on either 1 or
0, and adjust.